% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.R
\name{naturalfactor}
\alias{naturalfactor}
\title{Natural Ordering Factor}
\usage{
naturalfactor(x, levels, ordered = TRUE, ...)
}
\arguments{
\item{x}{a character vector.}

\item{levels}{a character vector whose elements might be appeared in \code{x}.}

\item{ordered}{logical flag that determines whether the factor is ordered.}

\item{...}{arguments that are passed to \code{factor} function.}
}
\description{
\code{naturalfactor} creates a factor with levels in natural order.
}

