\name{naturalsort-package}
\alias{naturalsort-package}
\docType{package}
\title{
naturalsort
}
\description{
The package provides functions related to natural ordering.
}
\details{
\tabular{ll}{
Package: \tab naturalsort\cr
Type: \tab Package\cr
Version: \tab 0.1.1\cr
Date: \tab 2013-10-03\cr
License: \tab BSD 3-clause License\cr
}
}
\keyword{ natural ordering }
\keyword{ natural sort }
