% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order.R, R/sort.R
\name{naturalorder}
\alias{naturalorder}
\alias{naturalsort}
\title{Natural Ordering Sort}
\usage{
naturalorder(text, decreasing = FALSE, na.last = TRUE)

naturalsort(text, decreasing = FALSE, na.last = NA)
}
\arguments{
\item{text}{a character vector to sort.}

\item{decreasing}{logical.}

\item{na.last}{logical. If \code{NA}, \code{NA}s will be removed of the result.}
}
\value{
For \code{naturalorder}, the results are indices of vector elements in natural order.
For \code{naturalsort}, the results are sorted vectors.
}
\description{
Natural ordering is a kind of alphanumerical ordering.
\code{naturalorder} returns the order of the argument character #' vector in human natural ascending or descending order.
\code{naturalsort} returns the sorted vector.
}
\examples{
text <- c("a-1.png", "a-2.png", "a-10.png")
print(sort(text))
print(naturalsort(text))

}

